+++ packages/nuomi-native/src/components/propTypes.js
@@ -3,9 +3,9 @@ import PropTypes from 'prop-types';
 export const NuomiNativeProps = {
   id: PropTypes.string,
   state: PropTypes.object,
-  reducer: PropTypes.objectOf(PropTypes.func),
-  action: PropTypes.objectOf(PropTypes.func),
-  getter: PropTypes.objectOf(PropTypes.func),
+  reducers: PropTypes.objectOf(PropTypes.func),
+  actions: PropTypes.objectOf(PropTypes.func),
+  getters: PropTypes.objectOf(PropTypes.func),
   render: PropTypes.func,
   onInit: PropTypes.func,
   onShow: PropTypes.func,